+++ tests/cmds/test_users.py
@@ -293,7 +293,10 @@ def test_add_user_role_raises_error_when_username_does_not_exist(
     ]
     result = runner.invoke(cli, command, obj=cli_state)
     assert result.exit_code == 1
-    assert "User 'not_a_username@example.com' does not exist." in result.output
+    assert (
+        "User 'not_a_username@example.com' does not exist or you do not have permissio to view them."
+        in result.output
+    )
 
 
 def test_remove_user_role_removes(
@@ -342,7 +345,10 @@ def test_remove_user_role_raises_error_when_username_does_not_exist(
     ]
     result = runner.invoke(cli, command, obj=cli_state)
     assert result.exit_code == 1
-    assert "User 'not_a_username@example.com' does not exist." in result.output
+    assert (
+        "User 'not_a_username@example.com' does not exist or you do not have permission to view them."
+        in result.output
+    )
 
 
 def test_update_user_calls_update_user_with_correct_parameters_when_only_some_are_passed(